+++ src/cyphers/CryptoSystem.scala
@@ -27,10 +27,10 @@ abstract class Key [T]
 
 abstract class CryptoSystem [T]
 {
-  def _encrypt (msg: Array[Byte], key: T, seed: Int): Array[Byte]
+  def _encrypt (msg: Array[Byte], key: T, seed: Long): Array[Byte]
   def _decrypt (msg: Array[Byte], key: T): Array[Byte]
   
-  def encrypt (msg: String , key: Key [T], seed: Int = 0): String =
+  def encrypt (msg: String , key: Key [T], seed: Long = 0): String =
     arrayToNetwork(_encrypt(hostToArray(msg), key.getPublic, seed))
   def decrypt (msg: String , key: Key [T]): String =
     arrayToHost(_decrypt(networkToArray(msg), key.getPrivate))
@@ -39,7 +39,7 @@ abstract class CryptoSystem [T]
 trait EncapsulatedCrypto
 {
   type T
-  def makeKey(seed: Int) : Key[T]
+  def makeKey(seed: Long) : Key[T]
   def encrypt(msg: String, key: Key[T]) : String
   def decrypt(msg: String, key: Key[T]) : String
   // TODO : information complémentaire
